adding styles to group page
