+++ static/style.css
@@ -1,7 +1,24 @@
 body {
   font-family: "helvetica", sans-serif;
-  margin: 16px;
+   width:100%
 }
+
+footer.fixed-bottom{
+  position: fixed;
+  height: 138px;
+  background: #FEFEFE ;
+  border-top: 1px solid rgb(206, 212, 218);
+  bottom: 0;
+  padding: 5px;
+  width: 100%;
+}
+
+.foot_padding{
+  height: 160px !important ;
+  width: 100%;
+
+}
+/*
 ul {
   list-style-type: none;
   padding-left: 0px;
@@ -60,17 +77,8 @@ h5.waitlist {
 #player_type li {
   display: inline;
   margin: 10px;
-}
-#add_submit {
-  height: 30px;
-  width: 70px;
-}
-#remove_submit {
-  height: 30px;
-  width: 70px;
-  margin-left: 20px;
-}
+} */
 #flashes {
   background-color: #faa;
   padding: 8px;
-}
+}
